+++ src/Tests/AzCli.fs
@@ -3,6 +3,7 @@ module AzCli
 open Expecto
 open Farmer
 open System
+
 let tests = testList "Azure CLI" [
     test "Can connect to Az CLI" {
         match Deploy.checkVersion Deploy.Az.MinimumVersion with
@@ -15,7 +16,7 @@ let tests = testList "Azure CLI" [
         Expect.equal result (Error "The following parameters are missing: p1.") ""
     }
 
-    test "Deploys and Deletes a resource group" {
+    test "Deploys and deletes a resource group" {
         let deployment = arm { location NorthEurope }
         let resourceGroupName = sprintf "farmer-integration-test-delete-%O" (Guid.NewGuid())
         printfn "Creating resource group %s..." resourceGroupName