+++ src/Tests/Storage.fs
@@ -10,12 +10,12 @@ open System
 
 /// Client instance needed to get the serializer settings.
 let client = new StorageManagementClient(Uri "http://management.azure.com", TokenCredentials "NotNullOrWhiteSpace")
-let tests = testList "Storage Tests" [
+let tests = testList "Storage" [
     test "Can create a basic storage account" {
         let resource =
             let account = storageAccount {
                 name "myStorage123~@"
-                sku Sku.PremiumLRS
+                sku StorageSku.Premium_LRS
             }
             arm { add_resource account }
             |> findAzureResources<StorageAccount> client.SerializationSettings